refactored code - Last review
