--- Week1/querry.js
@@ -1,1 +1,0 @@
-
